Move snapshot tests to a separate test class
